# server/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GapMiner API
  description: |
    GapMiner is a research gap discovery platform. This API covers paper management,
    gap extraction, AI analysis, collections, search, community features, analytics,
    and feature gating.
  version: 1.0.0
  contact:
    name: GapMiner Support
    url: https://github.com/anomalyco/gapminer
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: https://api.gapminer.com/v1
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /papers:
    get:
      tags: [Papers]
      summary: List user papers with search
      operationId: listPapers
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: q
          in: query
          schema: { type: string }
          description: Full-text search query
        - name: venue
          in: query
          schema: { type: string }
        - name: year
          in: query
          schema: { type: integer }
        - name: sort
          in: query
          schema: { type: string, enum: [created_at, title, year, citation_count], default: created_at }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        "200":
          description: Paginated list of papers
          content:
            application/json:
              schema:
                type: object
                properties:
                  papers:
                    type: array
                    items:
                      $ref: "#/components/schemas/Paper"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
    post:
      tags: [Papers]
      summary: Create a new paper
      operationId: createPaper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaperRequest"
      responses:
        "201":
          description: Paper created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paper"
        "400":
          $ref: "#/components/responses/ValidationError"
  /papers/{id}:
    parameters:
      - $ref: "#/components/parameters/PaperId"
    get:
      tags: [Papers]
      summary: Get paper details with gaps
      operationId: getPaper
      responses:
        "200":
          description: Paper with associated gaps
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Paper"
                  - type: object
                    properties:
                      gaps:
                        type: array
                        items:
                          $ref: "#/components/schemas/Gap"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Papers]
      summary: Delete a paper and its gaps
      operationId: deletePaper
      responses:
        "200":
          description: Paper deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Gaps ────────────────────────────────────────────────────────────────────
...
  /collections/{id}/papers:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Collections]
      summary: Add paper to collection
      operationId: addPaperToCollection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paperId]
              properties:
                paperId:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Paper added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login or /auth/register
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    PaperId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    TokenPair:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
...
    CreatePaperRequest:
      type: object
      required: [url, title]
      properties:
        url: { type: string, format: uri }
        title: { type: string, minLength: 1, maxLength: 500 }
        abstract: { type: string }
        authors: { type: array, items: { type: string } }
        venue: { type: string }
        year: { type: integer, minimum: 1900, maximum: 2100 }
        content: { type: string }
        source: { type: string }
        published_at: { type: string, format: date-time }
        metadata: { type: object, additionalProperties: true }
    Paper:
      type: object
      properties:
        id: { type: string, format: uuid }
        url: { type: string }
        title: { type: string }
        abstract: { type: string }
        authors: { type: array, items: { type: string } }
        venue: { type: string }
        year: { type: integer }
        citation_count: { type: integer }
        metadata: { type: object }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        gap_count: { type: integer }

    # ── Gaps ──────────────────────────────────────────────────────────────────
...
    Gap:
      type: object
      properties:
        id: { type: string, format: uuid }
        paper_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        problem: { type: string }
        type: { type: string }
        confidence: { type: number }
        impact_score: { type: string }
        difficulty: { type: string }
        assumptions: { type: array, items: { type: string } }
        failures: { type: array, items: { type: string } }
        dataset_gaps: { type: array, items: { type: string } }
        evaluation_critique: { type: string }
        upvotes: { type: integer }
        is_resolved: { type: boolean }
        created_at: { type: string, format: date-time }
        paper_title: { type: string }
        paper_url: { type: string }
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
    MessageResponse:
      type: object
      properties:
        message: { type: string }
...
    Error:
      type: object
      properties:
        error: { type: string }
        details: { type: array, items: { type: object } }
        code: { type: string }
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...